+++ ext/http/service.rs
@@ -182,6 +182,25 @@ impl Drop for HttpRequestBodyAutocloser {
   }
 }
 
+#[allow(clippy::collapsible_if)] // for logic clarity
+fn validate_request(req: &Request) -> bool {
+  if req.uri() == "*" {
+    if req.method() != http::Method::OPTIONS {
+      return false;
+    }
+  } else if req.uri().path().is_empty() {
+    if req.method() != http::Method::CONNECT {
+      return false;
+    }
+  }
+
+  if req.method() == http::Method::CONNECT && req.uri().authority().is_none() {
+    return false;
+  }
+
+  true
+}
+
 pub(crate) async fn handle_request(
   request: Request,
   request_info: HttpConnectionProperties,
@@ -189,6 +208,13 @@ pub(crate) async fn handle_request(
   tx: tokio::sync::mpsc::Sender<Rc<HttpRecord>>,
   legacy_abort: bool,
 ) -> Result<Response, hyper_v014::Error> {
+  if !validate_request(&request) {
+    let mut response = Response::new(HttpRecordResponse(None));
+    *response.version_mut() = request.version();
+    *response.status_mut() = http::StatusCode::BAD_REQUEST;
+    return Ok(response);
+  }
+
   let otel_info = if let Some(otel) = deno_telemetry::OTEL_GLOBALS
     .get()
     .filter(|o| o.has_metrics())
@@ -510,7 +536,7 @@ impl HttpRecord {
   /// Take the response.
   fn into_response(self: Rc<Self>) -> Response {
     let parts = self.self_mut().response_parts.take().unwrap();
-    let body = HttpRecordResponse(ManuallyDrop::new(self));
+    let body = HttpRecordResponse(Some(ManuallyDrop::new(self)));
     Response::from_parts(parts, body)
   }
 
@@ -592,8 +618,10 @@ impl HttpRecord {
   }
 }
 
+// `None` variant used when no body is present, for example
+// when we want to return a synthetic 400 for invalid requests.
 #[repr(transparent)]
-pub struct HttpRecordResponse(ManuallyDrop<Rc<HttpRecord>>);
+pub struct HttpRecordResponse(Option<ManuallyDrop<Rc<HttpRecord>>>);
 
 impl Body for HttpRecordResponse {
   type Data = BufView;
@@ -604,7 +632,9 @@ impl Body for HttpRecordResponse {
     cx: &mut Context<'_>,
   ) -> Poll<Option<Result<Frame<Self::Data>, Self::Error>>> {
     use crate::response_body::PollFrame;
-    let record = &self.0;
+    let Some(record) = &self.0 else {
+      return Poll::Ready(None);
+    };
 
     let res = loop {
       let mut inner = record.self_mut();
@@ -648,7 +678,7 @@ impl Body for HttpRecordResponse {
     }
 
     if let ResponseStreamResult::NonEmptyBuf(buf) = &res {
-      let mut http = self.0 .0.borrow_mut();
+      let mut http = record.0.borrow_mut();
       if let Some(otel_info) = &mut http.as_mut().unwrap().otel_info {
         if let Some(response_size) = &mut otel_info.response_size {
           *response_size += buf.len() as u64;
@@ -660,7 +690,10 @@ impl Body for HttpRecordResponse {
   }
 
   fn is_end_stream(&self) -> bool {
-    let inner = self.0.self_ref();
+    let Some(record) = &self.0 else {
+      return true;
+    };
+    let inner = record.self_ref();
     matches!(
       inner.response_body,
       ResponseBytesInner::Done | ResponseBytesInner::Empty
@@ -668,16 +701,22 @@ impl Body for HttpRecordResponse {
   }
 
   fn size_hint(&self) -> SizeHint {
+    let Some(record) = &self.0 else {
+      return SizeHint::with_exact(0);
+    };
     // The size hint currently only used in the case where it is exact bounds in hyper, but we'll pass it through
     // anyways just in case hyper needs it.
-    self.0.self_ref().response_body.size_hint()
+    record.self_ref().response_body.size_hint()
   }
 }
 
 impl Drop for HttpRecordResponse {
   fn drop(&mut self) {
+    let Some(record) = &mut self.0 else {
+      return;
+    };
     // SAFETY: this ManuallyDrop is not used again.
-    let record = unsafe { ManuallyDrop::take(&mut self.0) };
+    let record = unsafe { ManuallyDrop::take(record) };
     http_trace!(record, "HttpRecordResponse::drop");
     record.finish();
   }